Fixed compilation on OSX 10.6
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@9253 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ tools/pbespgsolve/ParityGame_pbes.h
@@ -28,8 +28,8 @@ void ParityGame::assign_pbes( mcrl2::pbes_system::pbes<Container> &pbes,
     verti begin = 0, end = 3;
     for (verti v = begin; v < end; ++v)
     {
-        std::set<unsigned int> deps = pgg.get_dependencies(v);
-        for ( std::set<unsigned int>::const_iterator it = deps.begin();
+        std::set<long unsigned int> deps = pgg.get_dependencies(v);
+        for ( std::set<long unsigned int>::const_iterator it = deps.begin();
                 it != deps.end(); ++it )
         {
             verti w = (verti)*it;
